Add support for absolute positioning
Unreal supports specifying the scale, position, or rotation as absolute, so add support for that as well. This requires the world transform itself to be propagated, not just the matrix derived from it. Since computing the matrix is noticably faster, I've implemented calculation of the WorldTransform as a lazy operation, so it will only be calculated for the few components that need it, rather than for the entire scene.