Add support for colors as diffuse textures
For games that use color maps, some materials will just be a base color with a normal map applied to it. This adds support for these scenarios.

This also adds support for a couple other fallback names that *something* usable as a diffuse texture may be getting passed in as. Color map is in this list rather than being added to CMaterialParams2.Diffuse because they aren't actually a diffuse texture, and instead are a greyscale map that gets multiplied by the diffuse color.